requirements.txt
aenum==3.1.15
appnope==0.1.4
asttokens==2.4.1
basemap==1.4.1
basemap-data==1.3.2
Cartopy==0.23.0
certifi==2024.8.30
cftime==1.6.4
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
executing==2.1.0
fonttools==4.53.1
geojson==3.1.0
geomet==1.1.0
geopandas==1.0.1
h5py==3.12.1
html2text==2024.2.26
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
jplephem==2.22
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
llvmlite==0.43.0
marshmallow==3.22.0
matplotlib==3.8.4
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
nest-asyncio==1.6.0
netCDF4==1.7.1.post2
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
packaging==23.2
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.3
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyogrio==0.9.0
pyparsing==3.1.4
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
seaborn==0.13.2
sentinelhub==3.10.3
sentinelsat==1.2.1
sgp4==2.23
shapely==2.0.6
six==1.16.0
skyfield==1.49
stack-data==0.6.3
tifffile==2024.8.30
tomli==2.0.1
tomli_w==1.0.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
utm==0.7.0
wcwidth==0.2.13
